cache_service/requirements.txt
Django==3.2.9
requests==2.26.0
redis==4.0.2
django-redis==5.0.0
whitenoise==5.3.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-dotenv==1.0.0